Simplify block queries in Crystal persistence.

+++ submerge-crystal/src/types/api/dto/block.rs
@@ -33,8 +33,8 @@ pub struct BlockQuery {
     pub max_block_number: Option<u64>,
     pub min_block_timestamp: Option<u64>,
     pub max_block_timestamp: Option<u64>,
-    pub min_spec_version: Option<u64>,
-    pub max_spec_version: Option<u64>,
+    pub min_spec_version: Option<u32>,
+    pub max_spec_version: Option<u32>,
     pub author: Option<AccountId>,
 }
